Avoid error when container::mapped_type is defined as void
Found a problem when trying to serialize this container since `mapped_type` was defined for both the set and map type. May be worth having a static assert to catch when someone else does this and give a more readable error.
https://github.com/martinus/unordered_dense/issues/48

Currently this is my solution so that the libraries function together.